requirements.txt
aiohttp==3.8.6
asyncpg==0.28.0
pydantic==2.4.2
SQLAlchemy==2.0.22
python-decouple==3.8
greenlet==3.0.0
google-cloud-storage==2.12.0
pyJWT==2.8.0
pytest==7.4.2
pytest-asyncio==0.21.1
aiokafka==0.8.1